requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.6
PyMySQL==1.1.0
PyYAML==6.0.1
requests==2.31.0
setuptools==68.2.2
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.1.0
wheel==0.41.2
